fix(build): resolve Next.js build errors on Vercel - add default exports to modal/BalanceSheet components - force-dynamic runtime for API/pages using headers/cookies/request.url - split GL page into ClientGL with ssr:false to avoid browser-only on SSR - rename next.config to mjs (ESM) - add ver comments
