+++ app/api/wholesale/products/statistics/route.ts
@@ -1,4 +1,7 @@
-// /api/wholesale/products/statistics/route.ts ver.1
+// ver.2 (2025-08-19 JST) - disable prerender; runtime=node; no revalidate
+export const dynamic = 'force-dynamic';
+export const runtime = 'nodejs';
+export const revalidate = 0;
 import { NextResponse } from 'next/server';
 import { createClient } from '@supabase/supabase-js';
